[#7] Refactor: Sort swift files by their function
